Move MemoryRunStorage to common util
